first public release of xANNA content and docs
Most of the changes from the internal, specific version has
been the removal of information in order to transform a specific
ANNA into a prototype (xANNA).

The pretty pictures are just that.

Some of the specific references (e.g. name and public key of the
registrar) are placeholders. They clearly need to be replaced
when the prototype site is specialized.